fix docker error due to inheritance
